refactor(html): move js/dynamic pages
